requirements.txt
aiocron==1.8
aiofiles==0.6.0
aiohttp==3.7.4.post0
aiosignal==1.2.0
aiosqlite==0.17.0
appdirs==1.4.4
async-generator==1.10
async-timeout==3.0.1
asyncio-extras==1.3.2
asyncprawcore==2.3.0
attrs==21.4.0
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.10
croniter==1.2.0
cssselect==1.1.0
discord.py==1.7.3
fake-useragent==0.1.11
feedparser==6.0.8
ffmpeg==1.4
frozenlist==1.3.0
idna==3.3
importlib-metadata==4.10.1
lxml==4.7.1
multidict==5.2.0
numpy==1.22.1
pandas==1.4.0
parse==1.19.0
pdfkit==1.0.0
praw==7.5.0
prawcore==2.3.0
pycparser==2.21
pyee==8.2.2
PyNaCl==1.5.0
pyppeteer==1.0.2
pyquery==1.4.3
python-dateutil==2.8.2
python-dotenv==0.18.0
pythonbible==0.6.1
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
requests-html==0.10.0
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.3.1
tqdm==4.62.3
typing_extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
update-checker==0.18.0
urllib3==1.26.8
w3lib==1.22.0
websocket-client==1.2.3
websockets==10.1
yahoo-fin==0.8.9.1
yarl==1.7.2
youtube-dl==2021.12.17
zipp==3.7.0
